ExESDB.GatewayWorker (ex_esdb v0.7.3)

GatewayWorker processes are started on each node in the cluster, and contain the implementation functions for the Gater.API.

Types

error()

@type error() :: term()

selector_type()

@type selector_type() :: String.t() | map()

store()

@type store() :: atom()

stream()

@type stream() :: String.t()

subscription_name()

@type subscription_name() :: String.t()

subscription_type()

@type subscription_type() :: :by_stream | :by_event_type | :by_event_pattern

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

gateway_worker_name(store_id)

start_link(opts)