ExESDB.StreamsReader (ex_esdb v0.7.1)

This module is responsible for reading events from a stream.

Functions

get_streams(store)

@spec get_streams(store :: atom()) :: {:ok, list()} | {:error, term()}

Returns a list of all streams in the store. ## Parameters

- `store` is the name of the store.

## Returns

- `{:ok, streams}`  if successful.

stream_events(store, stream_id, start_version, count, direction \\ :forward)

@spec stream_events(
  store :: atom(),
  stream_id :: any(),
  start_version :: integer(),
  count :: integer(),
  direction :: :forward | :backward
) :: {:ok, Enumerable.t()} | {:error, term()}

Streams events from stream in batches of count events, in a direction.

worker_id(store, stream_id)